refactor : 목적에 맞게 redis 서비스로 이동. 로깅 방법 변경
