+++ src/main/java/com/seowonn/mymap/service/RedisService.java
@@ -1,0 +1,9 @@
+package com.seowonn.mymap.service;
+
+public interface RedisService {
+  String getData(String key);
+
+  void setDataExpire(String key, String value, long duration);
+
+  void deleteData(String key);
+}
